Rename commands to orders
