--- examples/demo/src/commands/index.js
@@ -1,10 +1,0 @@
-import CommandIcon from '@material-ui/icons/AttachMoney';
-
-import CommandList from './CommandList';
-import CommandEdit from './CommandEdit';
-
-export default {
-    list: CommandList,
-    edit: CommandEdit,
-    icon: CommandIcon,
-};
